+++ src/Network/Network/Models/AzureFirewallPolicy/PSAzureFirewallPolicyApplicationRule.cs
@@ -35,6 +35,16 @@ namespace Microsoft.Azure.Commands.Network.Models
 
         [JsonProperty(Order = 7, PropertyName = "sourceIpGroups")]
         public List<string> SourceIpGroups { get; set; }
+        
+        [JsonProperty(Order = 8, PropertyName = "webCategories")]
+        public List<string> WebCategories { get; set; }
+
+        [JsonProperty(Order = 9, PropertyName = "targetUrls")]
+        public List<string> TargetUrls { get; set; }
+
+        [JsonProperty(Order = 10, PropertyName = "terminateTLS")]
+        public bool? TerminateTLS { get; set; }
+
 
         [JsonIgnore]
         public string ProtocolsText
@@ -54,7 +64,7 @@ namespace Microsoft.Azure.Commands.Network.Models
             get { return JsonConvert.SerializeObject(SourceIpGroups, Formatting.Indented); }
         }
 
-            [JsonIgnore]
+        [JsonIgnore]
         public string TargetFqdnsText
         {
             get { return JsonConvert.SerializeObject(TargetFqdns, Formatting.Indented); }
@@ -64,6 +74,18 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string FqdnTagsText
         {
             get { return JsonConvert.SerializeObject(FqdnTags, Formatting.Indented); }
+        }
+
+        [JsonIgnore]
+        public string WebCategoriesText
+        {
+            get { return JsonConvert.SerializeObject(WebCategories, Formatting.Indented); }
+        }
+
+        [JsonIgnore]
+        public string TargetUrlsText
+        {
+            get { return JsonConvert.SerializeObject(TargetUrls, Formatting.Indented); }
         }
 
         public void AddProtocol(string protocolType, uint port = 0)