+++ src/PostgreSql/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Add parameter MinimalTlsVersion
+* First version of flexible server
 
 ## Version 0.2.0
 * New/Update-AzPostgreSqlFirewallRule create a default name with time stamp when no name is passed (#12738)